chat_service/requirements.txt
# FastAPI & Server
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# Database & Storage
trino==0.328
qdrant-client==1.9.1
pymysql==1.1.0
SQLAlchemy==2.0.25

# ML & Embeddings
sentence-transformers==3.0.1
openai>=1.40.0
google-generativeai>=0.7.0

# Utils
numpy==1.24.3
pandas==2.0.3
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.5.3

# SQL Parsing & Security
sqlglot==24.0.0

